ローマ字対応表の関数をモジュール化
